+++ src/dry_foundation/testing/helpers/__init__.py
@@ -1,6 +1,6 @@
 """Helper tools to assist in testing applications."""
 
-from .database import TestHandler
+from .database import TestRepository
 from .routes import TestRoutes, unit_test_case
 
-__all__ = ["TestHandler", "TestRoutes", "unit_test_case"]
+__all__ = ["TestRepository", "TestRoutes", "unit_test_case"]